+++ tools/update_lint_rules/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Follow the [Keep a Changelog] format.
 
+## 4.0.0 - 2025-04-22
+
+### Changed
+
+- Migrated source of [rules.json] to [message.yaml].
+  - Also removed from [yumemi-inc/flutter-yumemi-lints]  since the value of `incompatibles` does not exist in `message.yaml`.
+  - [Status] in `Lint rules` has been reflected and status or maturity level has been activated.
+
 ## 3.1.0 - 2025-01-27
 
 ### Added
@@ -31,3 +39,11 @@ Follow the [Keep a Changelog] format.
 <!-- Links -->
 
 [Keep a Changelog]: https://keepachangelog.com
+
+[rules.json]: https://github.com/dart-lang/sdk/blob/ae6da8b926f208bf87d2e11375be5c611c27ee1b/pkg/linter/tool/machine/rules.json
+
+[message.yaml]: https://github.com/dart-lang/sdk/blob/ae6da8b926f208bf87d2e11375be5c611c27ee1b/pkg/linter/messages.yaml
+
+[Status]: https://dart.dev/tools/linter-rules#status
+
+[yumemi-inc/flutter-yumemi-lints]: https://github.com/yumemi-inc/flutter-yumemi-lints